+++ Assets/OxGKit/TweenSystem/Scripts/Runtime/DoTweenAnimEvent.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using DG.Tweening;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
-using DG.Tweening;
 
 namespace OxGKit.TweenSystem
 {
@@ -20,16 +21,28 @@ namespace OxGKit.TweenSystem
 
         private bool _lastTrigger = false;
 
+        /// <summary>
+        /// Reset the trigger flag
+        /// </summary>
         public void ResetTrigger()
         {
             this._lastTrigger = false;
         }
 
+        /// <summary>
+        /// Set the PlayMode
+        /// </summary>
+        /// <param name="playMode"></param>
         public void SetPlayMode(PlayMode playMode)
         {
             this.playMode = playMode;
         }
 
+        /// <summary>
+        /// Add a tween to the tweens list
+        /// </summary>
+        /// <param name="doTweenAnims"></param>
+        /// <returns></returns>
         public DoTweenAnimEvent AddDoTweenAnim(params DoTweenAnim[] doTweenAnims)
         {
             if (doTweenAnims != null && doTweenAnims.Length > 0)
@@ -40,6 +53,10 @@ namespace OxGKit.TweenSystem
             return this;
         }
 
+        /// <summary>
+        /// Remove a specific tween from the tweens list
+        /// </summary>
+        /// <param name="doTweenAnim"></param>
         public void RemoveDoTweenAnim(DoTweenAnim doTweenAnim)
         {
             if (this.doTweenAnims.Count == 0) return;
@@ -50,37 +67,91 @@ namespace OxGKit.TweenSystem
             }
         }
 
+        /// <summary>
+        /// Clear the tweens list
+        /// </summary>
         public void ClearDoTweenAnims()
         {
             this.doTweenAnims.Clear();
         }
 
+        /// <summary>
+        /// Plays in normal order
+        /// </summary>
         public void PlayNormal()
         {
             this.PlayTweens(true, null);
         }
 
+        /// <summary>
+        /// Plays in normal order
+        /// </summary>
+        /// <param name="endCallback"></param>
+        public void PlayNormal(TweenCallback endCallback = null)
+        {
+            this.PlayTweens(true, endCallback);
+        }
+
+        /// <summary>
+        /// Plays in reverse order
+        /// </summary>
         public void PlayReverse()
         {
             this.PlayTweens(false, null);
         }
 
+        /// <summary>
+        /// Plays in reverse order
+        /// </summary>
+        /// <param name="endCallback"></param>
+        public void PlayReverse(TweenCallback endCallback = null)
+        {
+            this.PlayTweens(false, endCallback);
+        }
+
+        /// <summary>
+        /// Automatically record the trigger: when the trigger is true, it will be in normal mode; when the trigger is false, it will be in reverse mode
+        /// </summary>
         public void PlayTrigger()
         {
             this._lastTrigger = !this._lastTrigger;
-            this.PlayTrigger(this._lastTrigger);
+            this.PlayTrigger(this._lastTrigger, null);
         }
 
+        /// <summary>
+        /// Automatically record the trigger: when the trigger is true, it will be in normal mode; when the trigger is false, it will be in reverse mode
+        /// </summary>
+        /// <param name="endCallback"></param>
+        public void PlayTrigger(TweenCallback endCallback = null)
+        {
+            this._lastTrigger = !this._lastTrigger;
+            this.PlayTrigger(this._lastTrigger, endCallback);
+        }
+
+        /// <summary>
+        /// When trigger is set to true, it will be in normal mode; when set to false, it will be in reverse mode
+        /// </summary>
+        /// <param name="trigger"></param>
+        /// <param name="endCallback"></param>
         public void PlayTrigger(bool trigger, TweenCallback endCallback = null)
         {
             this.PlayTweens(trigger, endCallback);
         }
 
+        /// <summary>
+        /// Allow the trigger to play only once. When the trigger is set to true, it will play in normal mode; when set to false, it will play in reverse mode
+        /// </summary>
+        /// <param name="trigger"></param>
         public void PlayTriggerOnce(bool trigger)
         {
             this.PlayTriggerOnce(trigger, null);
         }
 
+        /// <summary>
+        /// Allow the trigger to play only once. When the trigger is set to true, it will play in normal mode; when set to false, it will play in reverse mode
+        /// </summary>
+        /// <param name="trigger"></param>
+        /// <param name="endCallback"></param>
         public void PlayTriggerOnce(bool trigger, TweenCallback endCallback = null)
         {
             // trigger = true
@@ -105,26 +176,50 @@ namespace OxGKit.TweenSystem
             {
                 case PlayMode.Parallel:
                     {
+                        // Durations list
+                        List<float> durations = new List<float>();
+
+                        // Seq tweens
                         var seq = DOTween.Sequence();
+
                         for (int i = 0; i < this.doTweenAnims.Count; i++)
                         {
                             int idx = i;
-                            seq.AppendCallback(() => this.doTweenAnims[idx]?.PlayTween(trigger));
+                            var tween = this.doTweenAnims[idx];
+                            seq.AppendCallback(() => tween?.PlayTween(trigger));
+
+                            // Create a list of the maximum durations for all tweens
+                            if (tween != null) durations.Add(tween.GetMaxDurationTween().duration);
                         }
+
+                        // Find the maximum value from the list to use as the interval time for the endCallback
+                        float maxDuration = durations.Aggregate((a, b) => a > b ? a : b);
+                        seq.AppendInterval(maxDuration);
+
+                        // Add endCallback to the end
                         if (endCallback != null) seq.AppendCallback(endCallback);
                         seq.AppendCallback(() => seq.Kill());
                     }
                     break;
                 case PlayMode.Sequence:
                     {
+                        // Seq tweens
                         var seq = DOTween.Sequence();
+
                         for (int i = 0; i < this.doTweenAnims.Count; i++)
                         {
                             int idx = i;
-                            float duration = (this.doTweenAnims[i] == null) ? 0f : this.doTweenAnims[i].GetMaxDurationTween().duration;
-                            seq.AppendCallback(() => this.doTweenAnims[idx]?.PlayTween(trigger));
+                            var tween = this.doTweenAnims[idx];
+
+                            // Get the maximum duration of the tween
+                            float duration = (tween != null) ? tween.GetMaxDurationTween().duration : 0f;
+                            seq.AppendCallback(() => tween?.PlayTween(trigger));
+
+                            // The time interval of the tween
                             seq.AppendInterval(duration);
                         }
+
+                        // Add endCallback to the end
                         if (endCallback != null) seq.AppendCallback(endCallback);
                         seq.AppendCallback(() => seq.Kill());
                     }