+++ Assets/OxGKit/TweenSystem/Scripts/CHANGELOG.md
@@ -1,7 +1,12 @@
 ## CHANGELOG
 
+## [1.0.1] - 2025-01-12
+- Added methods with an endCallback parameter to DoTweenAnimEvent.
+- Modified DoTweenAnim to allow editing in Prefab Isolation mode during Play Mode.
+- Fixed issue with DoTweenAnimEvent's parallel mode not correctly calling the endCallback.
+
 ## [1.0.0] - 2024-02-02
-- Stabled
+- Stabled.
 
 ## [0.0.4-preivew] - 2023-07-23
 - Added preview feature for DoTweenAnime.